// app.client/src/api/models/InventoryItemInfoByCodeVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InventoryItemInfoByCodeVM = {
    inventoryItemId?: string | null;
    ownerBarcodeItemNum?: string | null;
    productDescription?: string | null;
    productNotes?: string | null;
    productUom?: string | null;
    productGrossWeightKg?: number;
    productPackage?: string | null;
    lastPricePaid?: number;
    isLotRequired?: boolean;
    isSnRequired?: boolean;
    isDateMfgRequired?: boolean;
    isDateExpRequired?: boolean;
    imageUrl?: string | null;
    locationCode?: string | null;
    warehouse?: string | null;
    warehouseCode?: string | null;
    warehouseID?: number;
    quantity?: number;
    lotNum?: string | null;
    dateMfg?: string | null;
    dateExp?: string | null;
    userName?: string | null;
    userFullName?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};